+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * New CLI command `--pyivf-make_version` with greater flexibility - you can specify all options via command line options (without needing a YAML file),
 or overwrite any of the parameters provided in the input file or distribution metadata. [DudeNr33]
 
+* Provide type information and add `py.typed` file. [DudeNr33]
+
 ### Internal
 
 * Switch to Poetry for packaging.